group and split up tests on Travis CI
