+++ test/compendium-bagit.js
