+++ test/compendium-delete.js
