+++ test/metadata-compendium.js
